requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==4.2.16
django-bootstrap5==25.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
zope.interface==7.2
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9